Allow spaces in the addSections option
I tried to use the option for adding extra section with the following value:

```
addSections: '{"documentation":{"prefix":"### Documentation updates:","labels":["documentation"]}}'
```

However, this result in the following error:

```
/usr/local/bundle/gems/github_changelog_generator-1.15.0/lib/github_changelog_generator/generator/entry.rb:71:in `rescue in parse_sections': There was a problem parsing your JSON string for sections: 767: unexpected token at '{"documentation":{"prefix":"###' (RuntimeError)
```

So basically, the option is split on the first space, causing an invalid value to be passed to the underlying script. Adding single brackets around should fix this problem.